[24.10.30] UserPageController 오타 수정
