+++ src/main/java/com/itwillbs/bookjuk/controller/main/UserPageController.java
@@ -1,0 +1,26 @@
+package com.itwillbs.bookjuk.controller.main;
+
+import com.itwillbs.bookjuk.util.SecurityUtil;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Slf4j
+@Controller
+public class UserPageController {
+    @GetMapping("/")
+    public String userMain(Model model) {
+        log.info("userMain");
+        log.info("userRole: {}", SecurityUtil.getUserRoles());
+        log.info("userName: {}", SecurityUtil.getUserName());
+        log.info("userNum: {}", SecurityUtil.getUserNum());
+        //회원의 이름을 가져오기!
+        model.addAttribute("userName", SecurityUtil.getUserName());
+
+        if (SecurityUtil.hasRole("ROLE_INACTIVE")){
+            return "redirect:/login/phone"; // 권한이 "ROLE_INACTIVE"라면 리다이렉트
+        }
+        return "userMain";
+    }
+}
